dataset/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pillow>=8.0.0
numpy>=1.20.0
matplotlib>=3.3.0
openai>=2.0.0
tqdm>=4.60.0

# Development dependencies (optional)
pytest>=6.0
black>=21.0
ruff>=0.1.0
mypy>=0.910

# Optional dependencies for enhanced functionality
datasets>=2.0.0
transformers>=4.20.0